Added Bryan's Functions
Still working on getARC but here's what I've done so far.
I think the rules state that the 10 kpi bonus for the person with the
most arcs is only transferred to another player if that player has MORE
arcs than the last person who had the bonus so I coded getMostArcs as
such.